moving common methods into one parent class
